Move generators up to cfl_common and some test helper functions
