+++ portal/tests/test_class.py
@@ -36,8 +36,10 @@
 # identified as the original program.
 from __future__ import absolute_import
 
+from common.tests.utils.classes import create_class_directly
+from portal.tests.utils.classes import create_class, transfer_class
+
 from .base_test import BaseTest
-from .utils.classes import create_class, create_class_directly, transfer_class
 from .utils.messages import (
     is_class_created_message_showing,
     is_class_nonempty_message_showing,
