+++ portal/tests/test_independent_student.py
@@ -41,10 +41,11 @@ import time
 from django.core import mail
 from selenium.webdriver.support.wait import WebDriverWait
 
-from portal.tests.utils.classes import create_class_directly
+from common.tests.utils.classes import create_class_directly
 from portal.tests.utils.organisation import create_organisation_directly
 from portal.tests.utils.student import create_school_student_directly
 from portal.tests.utils.teacher import signup_teacher_directly
+
 from .base_test import BaseTest
 from .pageObjects.portal.home_page import HomePage
 from .utils import email as email_utils
@@ -57,8 +58,8 @@ from .utils.messages import (
 from .utils.student import (
     create_independent_student,
     create_independent_student_directly,
-    submit_independent_student_signup_form,
     signup_duplicate_independent_student_fail,
+    submit_independent_student_signup_form,
 )
