+++ portal/tests/test_organisation.py
@@ -37,13 +37,15 @@
 from __future__ import absolute_import
 
 from builtins import range
+
 from django.core import mail
 
+from common.tests.utils.classes import create_class_directly
 from portal.tests.pageObjects.portal.base_page import BasePage
 from portal.tests.pageObjects.portal.home_page import HomePage
+
 from .base_test import BaseTest
 from .utils import email as email_utils
-from .utils.classes import create_class_directly
 from .utils.messages import is_organisation_created_message_showing
 from .utils.organisation import (
     create_organisation,
@@ -51,7 +53,7 @@ from .utils.organisation import (
     join_teacher_to_organisation,
 )
 from .utils.student import create_school_student_directly
-from .utils.teacher import signup_teacher_directly, generate_details
+from .utils.teacher import generate_details, signup_teacher_directly
 
 
 class TestOrganisation(BaseTest, BasePage):
