+++ portal/tests/test_school_student.py
@@ -36,9 +36,10 @@
 # identified as the original program.
 from __future__ import absolute_import
 
+from common.tests.utils.classes import create_class_directly
 from portal.tests.pageObjects.portal.home_page import HomePage
+
 from .base_test import BaseTest
-from .utils.classes import create_class_directly
 from .utils.messages import is_student_details_updated_message_showing
 from .utils.organisation import create_organisation_directly
 from .utils.student import create_school_student_directly
