+++ portal/tests/test_teacher_student.py
@@ -38,16 +38,17 @@ from __future__ import absolute_import
 
 import time
 
+from common.tests.utils.classes import create_class_directly
 from portal.tests.pageObjects.portal.home_page import HomePage
+
 from .base_test import BaseTest
-from .utils.classes import create_class_directly
 from .utils.organisation import (
     create_organisation_directly,
     join_teacher_to_organisation,
 )
 from .utils.student import (
-    create_school_student,
     create_many_school_students,
+    create_school_student,
     create_school_student_directly,
 )
 from .utils.teacher import signup_teacher_directly
