+++ portal/tests/utils/classes.py
@@ -35,34 +35,8 @@
 # program; modified versions of the program must be marked as such and not
 # identified as the original program.
 
-from common.models import Teacher, Class
 
-from portal.helpers.generators import generate_access_code
-
-
-def generate_details():
-    name = "Class %d" % generate_details.next_id
-    accesss_code = generate_access_code()
-
-    generate_details.next_id += 1
-
-    return name, accesss_code
-
-
-generate_details.next_id = 1
-
-
-def create_class_directly(teacher_email, class_name=None):
-    name, access_code = generate_details()
-
-    if class_name is not None:
-        name = class_name
-
-    teacher = Teacher.objects.get(new_user__email=teacher_email)
-
-    klass = Class.objects.create(name=name, access_code=access_code, teacher=teacher)
-
-    return klass, name, access_code
+from common.tests.utils.classes import generate_details
 
 
 def create_class(page):
